requirements.txt
asgiref==3.8.1
Django>=4.2,<5.0
django-cors-headers>=3.10.0
djangorestframework>=3.13.0
djangorestframework_simplejwt>=5.4.0
gunicorn
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
django-filter>=2.4.0
